refactor(dashboard-tabs): move mobile select variant into DashboardTabsList
Make the mobile dropdown opt-in via a `mobileVariant: 'tabs' | 'select'`
prop so it can be reused by other tab lists. Sandbox details layout now
consumes DashboardTabsList again instead of inlining the tab and select
primitives.
